аплоад картинки через http://localhost:8080/addProduct
http://localhost:8080/product

+++ src/main/java/com/example/springlesson3/interfaces/CategoryRepository.java
@@ -7,7 +7,7 @@ import java.util.List;
 
 public interface CategoryRepository {
 
-    Category get (Long id);
+    Category get (int id);
 
     List<Category> getAll();
